Extract init-creds and format commands from cli/index.ts
