+++ cli/BUILD
@@ -9,9 +9,11 @@ package(default_visibility = ["//visibility:public"])
 ts_library(
     name = "cli",
     srcs = [
+        "commands/format_command.ts",
         "commands/help_command.ts",
         "commands/index.ts",
         "commands/init_command.ts",
+        "commands/init_creds_command.ts",
         "commands/install_command.ts",
         "common_options.ts",
         "console.ts",
