+++ cli/commands/index.ts
@@ -1,3 +1,5 @@
+export { formatCommand } from "df/cli/commands/format_command";
 export { helpCommand } from "df/cli/commands/help_command";
-export { initCommand, icebergOption } from "df/cli/commands/init_command";
+export { initCommand } from "df/cli/commands/init_command";
+export { initCredsCommand } from "df/cli/commands/init_creds_command";
 export { installCommand } from "df/cli/commands/install_command";
